requirements.txt
# Multi-Agent Research System Dependencies
# ==========================================

# PDF Processing
PyMuPDF>=1.24.0

# LangChain Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Vector Database
faiss-cpu>=1.8.0

# Embeddings
sentence-transformers>=3.0.0

# Database
sqlalchemy>=2.0.0

# LLM Providers
langchain-openai>=0.2.0          # OpenRouter + OpenAI (required)
langchain-google-genai>=2.0.0    # Google Gemini (Free tier available)
# ollama                          # Ollama (Local, Free)

# UI
streamlit>=1.38.0
rich>=13.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
tabulate>=0.9.0
